Package DCC26 notebook helpers inside engiopt
