+++ engiopt/workshops/__init__.py
@@ -1,0 +1,1 @@
+"""Workshop helpers and assets shipped with EngiOpt."""
