+++ engiopt/workshops/dcc26/__init__.py
@@ -1,0 +1,3 @@
+"""DCC26 workshop utilities packaged with EngiOpt."""
+
+from .notebook_helpers import *  # noqa: F401,F403
